+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - `APIServerConfig`:
   - Added properties: `cacheStaticFilesResponses`, `staticFilesCacheMaxMemorySize` and `staticFilesCacheMaxContentLength`.
+- `APIConfig`:
+  - `getAs` and `getPath`: also tries to parse the value to the returned type.
 
 ## 1.5.11